add space for do command in app list

+++ Helpers/swda-deploy.sh
@@ -18,9 +18,15 @@
 
 
 cd /Library/Application\ Support/Purple/
+rm -rf _MACOSX
 curl -s -o /Library/Application\ Support/Purple/swda.zip https://raw.githubusercontent.com/PurpleComputing/mdmscripts/main/Helpers/swda.zip
 unzip swda.zip
 sleep 3
-rm swda.zip
+rm swda.zip _MACOSX
+
 xattr -r -d com.apple.quarantine /Library/Application\ Support/Purple/swda
-ln -s /Library/Application\ Support/Purple/swda /usr/local/bin/swda
+
+if [ ! -e "/usr/local/bin/swda" ]; then
+	ln -s /Library/Application\ Support/Purple/swda /usr/local/bin/swda
+fi
+
